+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Treat true/false as YES/NO when assigned as strings for xboolean filter
 * Support infinite ranges for date, datetime and integer filters
+* Treat `ActiveRecord::Result` class as `Array` driver
 
 ## 1.8.1