+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+## 1.8.2
+
 * Treat true/false as YES/NO when assigned as strings for xboolean filter.
 * Support infinite ranges for date, datetime and integer filters.
 * Treat `ActiveRecord::Result` class as `Array` driver.